feat: Update Visit: change type of priority from byte to int

+++ src/main/java/org/youcode/waitingroom/visit/application/dto/VisitRequestDTO.java
@@ -27,7 +27,7 @@ public record VisitRequestDTO(
         @NotBlank
         Status status,
 
-        Byte priority,
+        int priority,
 
         @NotNull
         int estimatedProcessingTime,
